Refactor validate license terms
